requirements.txt
Flask
pymongo
bcrypt
python-dotenv
Flask-Session  # For server-side sessions